文件大小使用long型表示

+++ src/main/java/me/zhenchuan/files/hdfs/App.java
@@ -33,7 +33,7 @@ public class App {
                 .setBaseWorkPath(properties.getProperty("work_path"))
                 .setFilenamePattern(properties.getProperty("file_pattern"))
                 .setGran(properties.getProperty("gran"))
-                .setMaxUploadSize(NumberUtils.toInt(properties.getProperty("max_upload_size"),200 * 1024 * 1024))
+                .setMaxUploadSize(NumberUtils.toLong(properties.getProperty("max_upload_size"),200 * 1024 * 1024))
                 .setTmpDir(properties.getProperty("tmp_dir","/tmp"));
 
         log.info("config:\n{}" , builder.toString());
